requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
pdfminer.six
httpx
tiktoken